+++ test/unit.test.ts
@@ -51,7 +51,7 @@ import './unit/cross-instance.test.js';
 import './unit/server.test.js';
 import './unit/plugin.test.js';
 import './unit/dexie-helper.test.js'
+import './unit/performance.test';
 import './unit/last.test.js';
 
 
-import './unit/performance.test';